TnLinter - fix to split books.
